chore: web file structure refactor
