+++ packages/web/.vscode/settings.json
@@ -8,6 +8,6 @@
   "files.readonlyInclude": {
     "**/routeTree.gen.ts": true
   },
-  "tailwindCSS.experimental.configFile": "./src/styles/app.css",
+  "tailwindCSS.experimental.configFile": "./src/styles.css",
   "tailwindCSS.classFunctions": ["tw", "clsx"]
 }
