+++ packages/web/src/client.tsx
@@ -4,6 +4,5 @@ import { StartClient } from "@tanstack/react-start";
 import { createRouter } from "./router";
 
 const router = createRouter();
-export type Router = typeof router;
 
 hydrateRoot(document, <StartClient router={router} />);
