+++ packages/web/src/components/form/buttons.tsx
@@ -1,7 +1,7 @@
-import { Button } from "./ui/button";
-import { useFormContext } from "./form";
+import { Button } from "../ui/button";
 import { useStore } from "@tanstack/react-form";
 import { cn } from "@/lib/utils";
+import { useFormContext } from ".";
 
 type CancelButtonProps = React.ComponentProps<typeof Button>;
