+++ packages/web/src/lib/participants.ts
@@ -10,17 +10,3 @@ export function getPayerFromParticipants(
 ) {
   return participants.find((p) => p.role === "payer")?.userId;
 }
-
-export function assertParticipantsHaveMembers(
-  participants: ParticipantWithMember[]
-) {
-  const truthy = participants.filter((p) => !!p.member) as Array<
-    ParticipantWithMember & { member: Member }
-  >;
-
-  if (truthy.length !== participants.length) {
-    throw new Error("All articipants must have members");
-  }
-
-  return truthy as Array<ParticipantWithMember & { member: Member }>;
-}
