+++ packages/web/src/lib/search-route.ts
@@ -12,7 +12,7 @@ export type UseSearchRoute = {
   state: () => unknown;
 };
 
-export type SearchRoute = {
+type SearchRoute = {
   useSearchRoute: () => UseSearchRoute;
 };
