+++ packages/web/src/pages/api/sync.push.ts
@@ -1,8 +1,28 @@
 import { json } from "@tanstack/react-start";
 import { createAPIFileRoute } from "@tanstack/react-start/api";
-import { createServerMutators, processor } from "@/server/mutators";
 import { AuthTokens } from "@/server/utils";
 import { authenticate } from "@/server/auth/core";
+import { connectionProvider } from "@rocicorp/zero/pg";
+import { schema } from "@blank/zero";
+import { PushProcessor } from "@rocicorp/zero/pg";
+import { OpenAuthToken } from "@blank/auth/subjects";
+import { Resource } from "sst";
+import postgres from "postgres";
+import { createClientMutators } from "@/lib/client-mutators";
+
+export const processor = new PushProcessor(
+  schema,
+  connectionProvider(postgres(Resource.Database.connection))
+);
+
+// whenever this grows, we can move it into server/server-mutators/*
+export function createServerMutators(auth: OpenAuthToken | undefined) {
+  const clientMutators = createClientMutators(auth);
+
+  return {
+    ...clientMutators,
+  };
+}
 
 export const APIRoute = createAPIFileRoute("/api/sync/push")({
   POST: async ({ request }) => {
